jss customButton and license page done
